Refactor symbolic mask propagation
