+++ nncf/tensorflow/pruning/base_algorithm.py
@@ -24,7 +24,7 @@ from nncf.common.graph.transformations.commands import TransformationPriority
 from nncf.common.pruning.clusterization import Cluster
 from nncf.common.pruning.clusterization import Clusterization
 from nncf.common.pruning.mask_propagation import MaskPropagationAlgorithm
-from nncf.common.pruning.pruning_node_selector import PruningNodeSelector
+from nncf.common.pruning.node_selector import PruningNodeSelector
 from nncf.common.pruning.statistics import PrunedLayerSummary
 from nncf.common.pruning.statistics import PrunedModelStatistics
 from nncf.common.pruning.structs import PrunedLayerInfoBase
